backend/requirements.txt
fastapi==0.109.0
pydantic==2.5.3
sentence-transformers>=2.7.0
umap-learn==0.5.5
python-multipart==0.0.6
uvicorn==0.27.0
numpy>=1.26.3
scikit-learn==1.4.0
